requirements.txt
asgiref==3.9.1
cffi==2.0.0
cryptography==46.0.1
Django==5.2.5
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-sendgrid-v5==1.3.0
MarkupSafe==3.0.2
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
python-dotenv==1.1.1
python-http-client==3.3.7
sendgrid==6.12.5
sqlparse==0.5.3
Werkzeug==3.1.3
